Fixed reading of variable-length GGEP items
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@4222 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/HostCache.cpp
@@ -1031,7 +1031,7 @@ void CHostCacheHost::Reset(IN_ADDR* pAddress)
 //////////////////////////////////////////////////////////////////////
 // CHostCacheHost update
 
-void CHostCacheHost::Update(WORD nPort, DWORD tSeen, LPCTSTR pszVendor, WORD nUptime)
+void CHostCacheHost::Update(WORD nPort, DWORD tSeen, LPCTSTR pszVendor, DWORD nUptime)
 {
 	m_nPort		= nPort;
 	m_tSeen		= tSeen > 1 ? tSeen : static_cast< DWORD >( time( NULL ) );
